split test_visualization.py into test_circuit_latex.py and test_utils.py
